# api/openapi.yaml
openapi: 3.0.3
info:
  title: Darkory
  version: 0.1.0
  description: |
    The canonical contract between Members and a Darkory Install. The CLI, the MCP server and the
    web app use only these operations; humans and agents call the same ones.

    **Versioning.** Paths are versioned (`/v1`). Changes within `/v1` are additive only.

    **Credentials.** Every operation except `getHealth`, `showLoginLink`, `redeemLoginLink` and
    `requestEmailSignIn` needs a credential: a bearer token (`dk_…`) together with a
    `Darkory-Session` header naming the Session making the request, or the browser cookie
    `darkory_session`. A bearer request without `Darkory-Session` is refused with
    `session_required`. A request with no credential is never a Member. A browser Session ends
    after a time unused and after a time in all (Install settings); a deactivated Member's
    credentials are all refused.

    **References.** Path parameters and request-body references take an id, or the natural key
    where one exists: a Task or Feature display key (`WEB-42`), a Team key (`WEB`), a Skill name
    (`qa`), a Member name. Tokens, Evidence and Sessions are referenced by id only. Responses
    always carry ids, plus display keys where they exist.

    **Retries.** Every write accepts an `Idempotency-Key` header. The first response is kept for
    24 hours per Member and key; a retry with the same key returns it unchanged. Reusing a key
...

...

...
    omitted.
  license:
    name: Apache-2.0
...
servers:
...
security:
  - bearerAuth: []
    sessionHeader: []
  - cookieAuth: []
...
paths:
...
  /v1/members:
    post:
      operationId: createMember
      tags: [members]
      summary: Create a Member (admin)
      description: "Errors: `forbidden`, `conflict` (name or email taken)."
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateMemberBody"}
      responses:
        "201":
          description: The new Member.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        default: {$ref: "#/components/responses/Error"}
    get:
      operationId: listMembers
      tags: [members]
      summary: List the Organisation's Members
      parameters:
        - name: team
          in: query
          description: Only Members of this Team.
          schema: {type: string}
        - name: kind
          in: query
          schema: {$ref: "#/components/schemas/MemberKind"}
      responses:
        "200":
          description: The Members, by name.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MemberList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/members/{member}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a Member with their Teams, Skills and Reporting line
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
      responses:
        "200":
          description: The Member.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MemberDetail"}
        default: {$ref: "#/components/responses/Error"}
    patch:
      operationId: updateMember
      tags: [members]
      summary: Change a Member's name, email or admin mark (admin)
      description: "Errors: `forbidden`, `conflict` (name or email taken; removing the last admin)."
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateMemberBody"}
      responses:
        "200":
          description: The Member.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        default: {$ref: "#/components/responses/Error"}
  /v1/members/{member}/deactivate:
    post:
      operationId: deactivateMember
      tags: [members]
      summary: Deactivate a Member (admin)
      description: |
        Revokes the Member's tokens, closes their Sessions and ends every Claim they hold, bound
        to a Session or to the Member, recording each in Activity. From then on every credential
        of theirs is refused, and no token or login link can be issued for them. The Member stays
        in the record, with everything they did. Deactivating a deactivated Member changes
        nothing. Errors: `forbidden` (not an admin, or deactivating yourself), `conflict` (the
        last active admin).
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "200":
          description: The deactivated Member.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        default: {$ref: "#/components/responses/Error"}
  /v1/members/{member}/reactivate:
    post:
      operationId: reactivateMember
      tags: [members]
      summary: Reactivate a deactivated Member (admin)
      description: |
        The Member may sign in and be issued tokens again. Tokens revoked and Sessions closed by
        the deactivation stay so. Reactivating an active Member changes nothing. Errors:
        `forbidden` (not an admin).
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "200":
          description: The Member.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        default: {$ref: "#/components/responses/Error"}
  /v1/members/{member}/skills/{skill}:
    put:
      operationId: grantSkill
      tags: [members]
      summary: Grant a Skill to a Member (admin)
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/SkillRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204":
          description: The Member has the Skill.
        default: {$ref: "#/components/responses/Error"}
    delete:
      operationId: revokeSkill
      tags: [members]
      summary: Take a Skill away from a Member (admin)
      description: Claims the Member holds under the Skill are not ended.
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/SkillRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204":
          description: The Member no longer has the Skill.
        default: {$ref: "#/components/responses/Error"}
  /v1/members/{member}/manager:
    put:
      operationId: setManager
      tags: [members]
      summary: Set the Member who directs this Member (admin)
      description: "Errors: `forbidden`, `cycle` (the Reporting line would loop)."
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SetManagerBody"}
      responses:
        "204":
          description: The Reporting line is set.
        default: {$ref: "#/components/responses/Error"}
    delete:
      operationId: clearManager
      tags: [members]
      summary: Remove the Member's Reporting line (admin)
      parameters:
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204":
          description: The Member reports to no one.
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- teams
  /v1/teams:
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a Team (admin)
      description: "Errors: `forbidden`, `conflict` (key or name taken)."
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateTeamBody"}
      responses:
        "201":
          description: The new Team.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Team"}
        default: {$ref: "#/components/responses/Error"}
    get:
      operationId: listTeams
      tags: [teams]
      summary: List the Organisation's Teams
      responses:
        "200":
          description: The Teams, by name.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TeamList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/teams/{team}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a Team and its Members
      parameters:
        - {$ref: "#/components/parameters/TeamRef"}
      responses:
        "200":
          description: The Team.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TeamDetail"}
        default: {$ref: "#/components/responses/Error"}
  /v1/teams/{team}/members/{member}:
    put:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a Member to a Team (admin)
      parameters:
        - {$ref: "#/components/parameters/TeamRef"}
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204":
          description: The Member is in the Team.
        default: {$ref: "#/components/responses/Error"}
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a Member from a Team (admin)
      description: Claims the Member holds on the Team's Tasks are not ended.
      parameters:
        - {$ref: "#/components/parameters/TeamRef"}
        - {$ref: "#/components/parameters/MemberRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204":
          description: The Member is not in the Team.
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- skills
  /v1/skills:
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a generic or company Skill (admin)
      description: |
        Publishes version 1 with the given body. A company Skill names the generic Skill it builds
        on. Errors: `forbidden`, `conflict` (name taken).
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSkillBody"}
      responses:
        "201":
          description: The new Skill and its first version.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillDetail"}
        default: {$ref: "#/components/responses/Error"}
    get:
      operationId: listSkills
      tags: [skills]
      summary: List the Organisation's Skills
      parameters:
        - name: kind
          in: query
          schema: {$ref: "#/components/schemas/SkillKind"}
      responses:
        "200":
          description: The Skills, by name.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/skills/{skill}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get a Skill with its current version
      parameters:
        - {$ref: "#/components/parameters/SkillRef"}
      responses:
        "200":
          description: The Skill.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillDetail"}
        default: {$ref: "#/components/responses/Error"}
  /v1/skills/{skill}/versions:
    get:
      operationId: listSkillVersions
      tags: [skills]
      summary: List a Skill's published versions
      parameters:
        - {$ref: "#/components/parameters/SkillRef"}
      responses:
        "200":
          description: The versions, newest first.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillVersionList"}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A Member token, `dk_…`. Sent together with `Darkory-Session`.
    sessionHeader:
      type: apiKey
      in: header
      name: Darkory-Session
      description: The id the running copy chose for its Session. Required with a bearer token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: darkory_session
      description: The browser Session, set by redeeming a login link.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        A key unique to this write. A retry with the same key returns the first response. It is 1
        to 255 printable ASCII characters, without spaces; any other is refused with `invalid`.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: "^[!-~]+$"}
    MemberRef:
      name: member
      in: path
      required: true
      description: Member id or name.
      schema: {type: string}
    TeamRef:
      name: team
      in: path
      required: true
      description: Team id or key, such as `WEB`.
      schema: {type: string}
    SkillRef:
      name: skill
      in: path
      required: true
      description: Skill id or name.
      schema: {type: string}
...
  responses:
    Error:
      description: The request failed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: {$ref: "#/components/schemas/ErrorCode"}
        message:
          type: string
          description: For people; may change between releases.
        details:
          type: object
          additionalProperties: true
          description: Extra facts about the failure, by code.
    ErrorCode:
      type: string
      description: |
        Stable reason for a failure, with its HTTP status:
        `invalid` 400 · `unauthenticated` 401 · `session_required` 401 · `forbidden` 403 ·
        `not_found` 404 · `conflict` 409 · `already_claimed` 409 · `not_takeable` 409 ·
        `not_holder` 409 · `ended` 409 · `tasks_open` 409 · `cycle` 409 · `proposal_stale` 409 ·
        `too_large` 413 · `idempotency_key_reused` 422 · `too_many_requests` 429 · `internal` 500 ·
        `not_implemented` 501.
      enum:
        - invalid
        - unauthenticated
        - session_required
        - forbidden
        - not_found
        - conflict
        - already_claimed
        - not_takeable
        - not_holder
        - ended
        - tasks_open
        - cycle
        - proposal_stale
        - too_large
        - idempotency_key_reused
        - too_many_requests
        - internal
        - not_implemented

    # ------------------------------------------------------------ install
...
    Member:
      type: object
      required: [id, name, kind, admin, created_at]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {$ref: "#/components/schemas/MemberKind"}
        email: {type: string}
        admin:
          type: boolean
          description: Admins create Members, Teams and Skills, set Reporting lines, and issue tokens and login links.
        manager_id:
          type: string
          description: The Member who directs this one. Absent when there is no Reporting line.
        created_at: {type: string, format: date-time}
        deactivated_at:
          type: string
          format: date-time
          description: When an admin deactivated the Member. Absent while the Member is active.
    MemberKind:
      type: string
      enum: [human, agent]
    MemberDetail:
      type: object
      required: [member, teams, skills, reports]
      properties:
        member: {$ref: "#/components/schemas/Member"}
        teams:
          type: array
          items: {$ref: "#/components/schemas/Team"}
        skills:
          type: array
          items: {$ref: "#/components/schemas/Skill"}
        reports:
          type: array
          description: The Members this Member directs.
          items: {$ref: "#/components/schemas/Member"}
    MemberList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Member"}
    CreateMemberBody:
      type: object
      required: [name, kind]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        kind: {$ref: "#/components/schemas/MemberKind"}
        email: {type: string, format: email}
        admin: {type: boolean}
    UpdateMemberBody:
      type: object
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        email: {type: string, format: email}
        admin: {type: boolean}
    SetManagerBody:
      type: object
      required: [manager]
      properties:
        manager:
          type: string
          description: Member id or name.

    # ------------------------------------------------------------ teams
    Team:
      type: object
      required: [id, key, name, created_at]
      properties:
        id: {type: string}
        key:
          type: string
          description: The prefix of the Team's display keys, such as `WEB` in `WEB-42`.
        name: {type: string}
        created_at: {type: string, format: date-time}
    TeamDetail:
      type: object
      required: [team, members]
      properties:
        team: {$ref: "#/components/schemas/Team"}
        members:
          type: array
          items: {$ref: "#/components/schemas/Member"}
    TeamList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Team"}
    CreateTeamBody:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          pattern: "^[A-Z][A-Z0-9]{1,9}$"
        name: {type: string, minLength: 1, maxLength: 100}

    # ------------------------------------------------------------ skills
    Skill:
      type: object
      required: [id, name, kind, builtin, current_version, created_at]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {$ref: "#/components/schemas/SkillKind"}
        base_skill_id:
          type: string
          description: The generic Skill a company Skill builds on.
        builtin:
          type: boolean
          description: True for `breakdown`, `retro` and `skill-review`, which Darkory relies on.
        current_version:
          type: integer
          format: int64
        created_at: {type: string, format: date-time}
    SkillKind:
      type: string
      enum: [generic, company]
    SkillVersion:
      type: object
      required: [skill_id, version, body, published_at]
      properties:
        skill_id: {type: string}
        version:
          type: integer
          format: int64
        body: {type: string}
        proposal_id:
          type: string
          description: The proposal this version was published from. Absent for version 1.
        published_by:
          type: string
          description: The Member who completed the review, or who created the Skill.
        published_at: {type: string, format: date-time}
    SkillDetail:
      type: object
      required: [skill, current]
      properties:
        skill: {$ref: "#/components/schemas/Skill"}
        current: {$ref: "#/components/schemas/SkillVersion"}
    SkillList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Skill"}
    SkillVersionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/SkillVersion"}
    CreateSkillBody:
      type: object
      required: [name, kind, body]
      properties:
        name:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
        kind: {$ref: "#/components/schemas/SkillKind"}
        base_skill:
          type: string
          description: Required for a company Skill. Id or name of a generic Skill.
        body:
          type: string
          description: The Skill's text, published as version 1.
...